Delivery-Tracking-System/Microservices/DeliveryTrackerv2/requirements.txt
Django>=4.0.4
djangorestframework>=3.13.1
pymongo>=4.1.1
dnspython>=2.2.1